adding image upload function

+++ src/components/Book.vue
@@ -8,57 +8,123 @@
             <div class="d-flex border">
               <div class="flex-grow-1 bg-white p-4">
                 <div v-if="currentBook" class="edit-form">
-                  <h4>Book</h4>
-                  <form>
-                    <div class="form-group">
-                      <label for="title">Title</label>
-                      <input
-                        type="text"
-                        class="form-control"
-                        id="title"
-                        v-model="currentBook.title"
-                      />
-                    </div>
-                    <div class="form-group">
-                      <label for="author">Author</label>
-                      <input
-                        type="text"
-                        class="form-control"
-                        id="author"
-                        v-model="currentBook.author"
-                      />
-                    </div>
-                    <div class="form-group">
-                      <label for="description">Description</label>
-                      <input
-                        type="text"
-                        class="form-control"
-                        id="description"
-                        v-model="currentBook.description"
-                      />
-                    </div>
+                  <b-form @submit.stop.prevent="onSubmit">
+                    <b-form-group
+                      id="title-edit-group"
+                      label="Title"
+                      label-for="title-edit"
+                      label-cols-sm="4"
+                      label-cols-lg="3"
+                      content-cols-sm
+                      content-cols-lg="7"
+                    >
+                      <b-form-input
+                        id="title-edit"
+                        name="title-edit"
+                        v-model="$v.currentBook.title.$model"
+                        :state="validateEditState('title')"
+                        aria-describedby="title-edit-live-feedback"
+                      ></b-form-input>
 
-                    <div class="form-group">
-                      <label><strong>Status:</strong></label>
-                      {{ currentBook.published ? "Published" : "Pending" }}
-                    </div>
+                      <b-form-invalid-feedback id="title-edit-live-feedback"
+                        >Title is a required field.</b-form-invalid-feedback
+                      >
+                    </b-form-group>
 
-                    <div class="form-group">
-                      <label>Cover image</label>
-                      <div v-if="!selectedImage">
-                        <input
-                          type="file"
-                          accept="image/*"
-                          @change="onFileChange(currentBook, $event)"
-                        />
+                    <b-form-group
+                      id="author-edit-group"
+                      label="Author"
+                      label-for="author-edit"
+                      label-cols-sm="4"
+                      label-cols-lg="3"
+                      content-cols-sm
+                      content-cols-lg="7"
+                    >
+                      <b-form-input
+                        id="author-edit"
+                        name="author-edit"
+                        v-model="$v.currentBook.author.$model"
+                        :state="validateEditState('author')"
+                        aria-describedby="author-edit-live-feedback"
+                      ></b-form-input>
+
+                      <b-form-invalid-feedback id="author-edit-live-feedback"
+                        >Author is a required field.</b-form-invalid-feedback
+                      >
+                    </b-form-group>
+
+                    <b-form-group
+                      id="description-edit-group"
+                      label="Description"
+                      label-for="description-edit"
+                      label-cols-sm="4"
+                      label-cols-lg="3"
+                      content-cols-sm
+                      content-cols-lg="7"
+                    >
+                      <b-form-textarea
+                        id="description-edit"
+                        name="description-edit"
+                        v-model="$v.currentBook.description.$model"
+                        :state="validateEditState('description')"
+                        aria-describedby="description-edit-live-feedback"
+                        rows="3"
+                        max-rows="6"
+                      ></b-form-textarea>
+
+                      <b-form-invalid-feedback
+                        id="description-edit-live-feedback"
+                        >Description is a required
+                        field.</b-form-invalid-feedback
+                      >
+                    </b-form-group>
+
+                    <b-form-group
+                      id="cover-edit-group"
+                      label="Cover Image"
+                      label-for="cover-edit"
+                      label-cols-sm="4"
+                      label-cols-lg="3"
+                      content-cols-sm
+                      content-cols-lg="7"
+                    >
+                      <b-form-file
+                        v-if="currentBook.image === null"
+                        id="cover-edit"
+                        name="cover-edit"
+                        v-model="$v.currentBook.image.$model"
+                        :state="validateEditState('image')"
+                        accept="image/jpeg, image/png, image/gif"
+                        placeholder="Choose an image"
+                        drop-placeholder="Drop file here..."
+                      ></b-form-file>
+                      <div class="image-container">
+                        <b-button
+                          v-if="hasImage"
+                          variant="danger"
+                          class="btn-sm"
+                          @click="clearImage"
+                          >Clear image</b-button
+                        >
+                        <b-img
+                          v-if="hasImage"
+                          :src="imageSrc"
+                          width="150"
+                          height="250"
+                          fluid
+                          block
+                          rounded
+                        ></b-img>
                       </div>
-                      <div v-else>
-                        <!-- <img :src="currentBook.data" /> -->
-                        <img v-bind:src="selectedImage" width="150px" height="200px" alt=""><br/>
-                        <button @click="removeImage(currentBook)">Remove image</button>
-                      </div>
-                    </div>
-                  </form>
+
+                      <b-form-invalid-feedback
+                        v-if="currentBook.image === null"
+                        id="cover-edit-live-feedback"
+                        >Cover image is a required
+                        field.</b-form-invalid-feedback
+                      >
+                    </b-form-group>
+                  </b-form>
 
                   <button
                     class="badge badge-primary mr-2"
@@ -104,43 +170,97 @@
 
 <script>
 import BookDataService from "../services/BookDataService";
+import { validationMixin } from "vuelidate";
+import { required } from "vuelidate/lib/validators";
+
+const base64Encode = (data) =>
+  new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.readAsDataURL(data);
+    reader.onload = () => resolve(reader.result);
+    reader.onerror = (error) => reject(error);
+  });
 
 export default {
+  mixins: [validationMixin],
   name: "book",
   data() {
     return {
       currentBook: null,
       message: "",
+      imageSrc: null,
     };
   },
+
+  validations: {
+    currentBook: {
+      title: { required },
+      author: { required },
+      description: { required },
+      image: { required },
+    },
+  },
+
+  computed: {
+    hasImage() {
+      if (!this.currentBook.image) return;
+
+      if (this.currentBook.image instanceof File) {
+        base64Encode(this.currentBook.image)
+          .then((value) => {
+            this.currentBook.image = value;
+            this.imageSrc = this.currentBook.image;
+          })
+          .catch(() => {
+            this.currentBook.image = null;
+            this.imageSrc = null;
+          });
+      } else return this.currentBook.image;
+    },
+
+    image() {
+      return this.currentBook.image;
+    },
+  },
+
+  watch: {
+    image(newValue, oldValue) {
+      if (newValue !== oldValue) {
+        if (newValue) {
+          base64Encode(newValue)
+            .then((value) => {
+              this.imageSrc = value;
+            })
+            .catch(() => {
+              this.imageSrc = null;
+            });
+        } else {
+          this.imageSrc = null;
+        }
+      }
+    },
+  },
+
   methods: {
+    clearImage() {
+      this.currentBook.image = null;
+      this.imageSrc = null;
+    },
+
     getBook(id) {
       BookDataService.get(id)
         .then((response) => {
           this.currentBook = response.data;
-          this.selectedImage = 'data:image/jpeg;base64,' + response.data.data;
+          this.imageSrc = "data:image/jpeg;base64," + response.data.image;
         })
         .catch((e) => {
           console.log(e);
         });
     },
 
-    onFileChange(book, e) {
-      var files = e.target.files || e.dataTransfer.files;
-      if (!files.length) return;
-      this.createImage(book, files[0]);
-    },
-    createImage(book, file) {
-      var reader = new FileReader();
-      reader.onload = (e) => {
-        book.data = e.target.result;
-      };
-      reader.readAsDataURL(file);
-    },
-    removeImage: function (book) {
-      book.data = false;
-      this.selectedImage = false;
-      this.currentBook.data = false;
+    validateEditState(title) {
+      const { $dirty, $error } = this.$v.currentBook[title];
+      return $dirty ? !$error : null;
     },
 
     updatePublished(status) {
@@ -150,7 +270,7 @@ export default {
         author: this.currentBook.author,
         description: this.currentBook.description,
         published: status,
-        data: this.currentBook.data,
+        image: this.currentBook.image,
       };
 
       BookDataService.update(this.currentBook.id, data)
@@ -164,7 +284,10 @@ export default {
     },
 
     updateBook() {
-      BookDataService.update(this.currentBook.id, this.currentBook)
+      BookDataService.update(
+        this.currentBook.id,
+        JSON.stringify(this.currentBook)
+      )
         .then((response) => {
           console.log(response.data);
           this.message = "The book was updated successfully!";
@@ -175,14 +298,16 @@ export default {
     },
 
     deleteBook() {
-      BookDataService.delete(this.currentBook.id)
-        .then((response) => {
-          console.log(response.data);
-          this.$router.push({ name: "Home" });
-        })
-        .catch((e) => {
-          console.log(e);
-        });
+      if (confirm("Are you sure you want to delete this item?")) {
+        BookDataService.delete(this.currentBook.id)
+          .then((response) => {
+            console.log(response.data);
+            this.$router.push({ name: "Home" });
+          })
+          .catch((e) => {
+            console.log(e);
+          });
+      }
     },
   },
   mounted() {
@@ -192,9 +317,14 @@ export default {
 };
 </script>
 
-<style>
-.edit-form {
-  max-width: 300px;
-  margin: auto;
+<style >
+.image-container {
+  position: relative;
+  max-width: 400px;
+  margin-top: 1rem;
+}
+
+.image-container .btn {
+  position: absolute;
 }
 </style>
